+++ examples/c/example_1/main.c
@@ -1,15 +1,12 @@
 
 #include <bdut/bdut.h>
 
-#include <stdio.h>
 
 int main(int argc, char* argv[])
 {
     BDUT_ASSERT_EQ(1, 2);
 
-    fprintf(stdout, "%s: ALL TESTS PASSED\n", argv[0]);
-
-    return 0;
+    return BDUT_TESTS_PASSED(argc, argv);
 }